Fix typing errors / cleanup
This commit fixes most of the introduced flow errors with express,
like introducing a `getContext` function to retrieve the context in
the appropriate target code.

Also it disables most of the ramda definitions, since the selectors
are so wrongly typed that flow gets huge performance hickups whenever
it tries to interpret our huge `R.compose` chains.

Maybe we will reenable them as soon after we refactored the selector code.