Amazeeio 914 add auth (#47)
* Refactor auth-server

This contains a few fixes:
- Make the couchdb auth database work again
- Use async / await in some instances
- Add timestamp to jwt.sign for different hashes

* Add README for auth-server

* Add basic token handling to Api

This will add basic token checking as a auth middleware.
There is no fine-grained credential checking via ssh-key yet.

* Add initial auth middleware logic

There are still many tasks open before this feature is complete:

- We need to discuss the final token behavior to retrieve the sshKey
  from
- We should create env variables for handling the secret key
- The credentials should be used in the GraphQL resolvers

* Add missing flow-typed stubs

* Add credential test data in local-hiera

* Add authorization layer

In this commit we introduce following changes:
- auth-server does add a role in the JWT token
- The "admin" role is allowed to access all data
- All resolvers now involve an authorization process

Currently, the api does not use the `auth-server/authenticate` interface
to verify that given JWT token is in its database. This will not
directly introduce any security problems, since the JWT token is still
verified in the authentication process.

After the token was verified, it will use the sshKey & role data to
process all accessible clients, sitegroups and sites. This is always
done beforehand. If the token contains an "admin" role, it will not
filter any entities, hence give access to all the data.

This commit also introduces fixes for the sshKey access. Previously, the
GraphQL schema could not yield valid values for the SshKey entity, since
`extractSshKeys` was returning plain ssh-key strings (and not SshKey
shaped objects). The fix considers the `api/keys` route as well to
preserve the original behavior.

* Add test for getCredentials w/ admin role

* Make auth-ssh compatible with auth-server / fix client

This commit will make it possible to add an admin token for the
auth-ssh services, so it can query all sshKeys from the API.

Also it introduces some small fixes to make the client compatible with
the API, like fixing the bearer token format.

* do not overwrite ENTRYPOINT so the upstream entrypoint system is used
Instead of overwriting the ENTRYPOINT we inject our entrypoint file as /amazeeio/entrypoints/99-envplate.sh which will be called by the amazeeio entrypoints handler automatically. We do that so all other entrypoint file (like the injector of env variables from .env files) is called)

* refactoring and adding of local AUTH_SSH_ADMIN_TOKEN
- not loading .env file from root, but instead adding .env.defaults directly to the services (which will be picked up by the amazeeio entrypoints system and exposed as env variables)
-  Generated an AUTH_SSH_ADMIN_TOKEN from the already used ssh private key for the git auth with admin rights and the jwtSecret 'super-secret-string'
- injecting default jwtSecret via .env.defaults files, you better change these in production!

* Extract JWT generation logic to another module

* Upgrade dependencies, format code

* Fix tests for new chalk version

* Format code

* Add to troubleshooting section

* Fix whitespace

* Fix Flow error by validating JWTSECRET env var

* Add /generate route for generating tokens via http

* Fix flow errors with Role type

* Add audience mechanics, JWTAUDIENCE env, remove obsolete routes

This will introduce following breaking changes for the API:
- An auth-token now needs to have an `aud` attribute, which matches
with the JWTAUDIENCE env variable in the api
- `user` role was renamed to `none` role
- `auth-server/generate` will now also return the payload used for
token generation

* Add missing routes.js file with /generate route

* Add missing flow-typed files