Make auth-ssh compatible with auth-server
There were a few adaptions missing to make the communication work
between auth-server and auth-ssh.

This commit also introduces small changes to the `/generate` endpoint,
which introduces a `verbose` parameter for JSON output.

If `verbose` is `false` or not set, it will just return the token
string as in our previous auth-server versions.