Add mariadb, add missing docker / Makefile changes
This commit introduces a MariaDB docker container, since there is only
limited support for postgres in the OCaml ecosystem (which we might
consider for our GraphQL endpoint later on).

Also since our core competence is MariaDB, it makes more sense to manage
our client & infrastructure data with the same technology.

(Also there are some smaller unrelated changes to the obsolete postgres
stuff, ... we keep postgres around in case we need it)