Modify sshKey update permission logic
Add:
Any authenticated user can add a sshKey.

Update:
A user can update a sshKey, if the key is either
assigned to a customer or a project the user
has access to.

Delete:
A user can delete a sshKey, if the key is either
assigned to a customer or a project the user has
access to.

Edge case: The user cannot yet delete any sshKey
which is not assigned to a project / customer the
user has access to.