Update lagoonize.md
Don't use "here" as link text; accessibility software usually has a "read all the links to me so I can choose the appropriate one to follow" action and hearing "here" is completely useless. Also, you are using Winnie-the-Pooh-style capitalization in A Few Places, which is Cute for a Children's Book, but looks Odd in This Context.

Lastly, I don't see anything about why you would need to “add `patches` directory if you choose drupal8-composer-mariadb”. Is this step still needed? If not, remove it. If it is, you would also need to update the `Dockerfile.cli` file to un-comment `COPY patches /app/patches`, right?