Update logs-forwarder configuration
The main change that seems to increase performance is switching from the
deprecated secure_forward plugin to the built-in forward plugin, which
supports TLS.

The advantages of the forward plugin are:

* it's maintained.
* support for fluentd workers so that input performance can be scaled
  easily.
* creates a new connection per chunk forward event.

On the last point: the old secure_forward plugin held onto its
connection which meant that sessions were "sticky" to a pod and
effectively all work was being sent to single forwarder pod.
With the forward plugin, the load balancer can do its job of spreading
work across all logs-fowarder pods.

The other changes include:
* flush threads are now set to 2 (x4 workers = 8)
* giving the output an @id so it is identifiable in logs
* not using a timekey, so chunk flushing is _only_ size dependent
* storing any tag/time in the same chunk to avoid lots of little chunks
* increasing the flush interval and chunk size
* dropping chunks on buffer overflow
* remove redundant index_name field in elasticsearch output
* make scheme configurable via environment variable (default http)