feat: add service-api client to keycloak
Previously only the api client had token exchange permissions.

This change adds another client "service-api", and gives it token
exchange permissions too.

Because of the way that the Keycloak API works, both clients had to be
set up before then adding the token exchange permissions to them both at
the same time. This required some refactoring of the api client setup
function, but the functionality of the api client should not have
changed.