feat: add attribute to groups which maps the group name to project IDs
This change adds an additional group annotation which consists of a JSON
blob containing a map of group name to project IDs. This makes it
possible, when getting the user token, to aggregate the annotations and
get all the groups and projects accessible to the user.

Crucially, it also makes it possible to discover through which group
membership a user has project access. This is important because a user
can have different roles in different groups which influences
authorization decisions.