fix: refactor changes to keycloak startup script
Don't modify the configure_api_client function, as this breaks startup
readiness logic in the API.

Instead assume that the functions are run in the required order, and
that global variables are set.

This is the assumption made in the rest of the script, and will be
reworked in future keycloak upgrade work.