fix: accept RSA SHA1 signatures in the ssh client for CI
OpenSSH has deprecated SHA1, and in 8.8 it was removed from the default
accepted signature algorithm list.

OpenSSH server implements signature algorithm negotiation. Go's SSH
server implementation does not. Since we use RSA keys in CI, the ssh
client uses those keys and because it can't negotiate an alternative
falls back to the default disallowed SHA1 algorithm, which causes the
connection to fail.

So for now to work around this problem we explicitly allow SHA1 in the
client.

Once signature negotiation is implemented in Go we can drop this patch.
See https://github.com/golang/crypto/pull/197.