fix: add custom dnsConfig to all cronjobs
This is an attempt to fix the problem where cronjobs intermittently fail
with getaddrinfo() error EAI_AGAIN. The problem appears to be that the
lookup times out due to dropped UDP packets.

The default for timeout is 5 seconds and attempts is 2 for both musl and
glibc. This change bumps the configured values to the maximum values for
musl: 60s timeout and 10 attempts.

For any custom cronjobs running glibc-based images the maximum values
are silently capped at half those values: 30s timeout and 5 attempts.

References:
https://git.musl-libc.org/cgit/musl/tree/src/network/resolvconf.c#n17
https://manpages.debian.org/buster/manpages/resolv.conf.5.en.html