Add some clarity to persistent storage
It wasn't clear to me that each container / service was able to have its own persistent volume, or was able to mount another services claim. I thought that there was one PV per Lagoon project. This was a misunderstanding on my part, but perhaps the proposed addition will make it clearer how PV works in Lagoon?