fix: avoid empty items in lagoon-project attribute list
The lagoon-projects group attribute in keycloak can sometimes end up
with a leading comma separator. This seems to happen if somehow the list
of project IDs has an empty string: "".

So this change ensures that the empty entries are stripped from keycloak
attributes by reusing the groupID parsing pipeline from elsewhere in the
file which already does this.